+++ dreamjob_frontend/src/App.js
@@ -8,7 +8,7 @@ import jobIndex from './routes/jobIndex';
 import RegisterUser from './routes/RegisterUser';
 import Home from './routes/Home';
 import jobDetails from './routes/jobDetails';
-import {updateJobs} from './actions';
+import {updateJobs} from './actions/actions';
 
 class App extends Component {
   constructor(props){
@@ -41,9 +41,9 @@ class App extends Component {
             <div>
               <Route exact path="/" component={Home}></Route>
               <Route exact path="/register" component={RegisterUser}></Route>
-              <Route exact path = '/addJob' component={addJob}></Route>
-              <Route exact path = '/jobIndex' component={jobIndex}></Route>
-              <Route exact path = '/jobDetails' component={jobDetails}></Route>
+              <Route exact path = '/add_job' component={addJob}></Route>
+              <Route exact path = '/job_index' component={jobIndex}></Route>
+              <Route exact path = '/job_details' component={jobDetails}></Route>
             </div>
           </Router>
         </div>
