--- dreamjob_frontend/src/Dispatcher.js
@@ -1,4 +1,0 @@
-import {Dispatcher} from 'flux'
-
-const dispatcher = new Dispatcher()
-export default dispatcher
