+++ dreamjob_frontend/src/routes/addJob.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 // import logo from '../logo.svg';
 import '../App.css';
-import {createJob} from '../actions'
+import {createJob} from '../actions/actions'
 import jobStore from '../stores/jobStore'
 
 class addJob extends Component {
@@ -24,7 +24,7 @@ class addJob extends Component {
 }
 
 redirect(){
-  this.props.history.push("/jobIndex")
+  this.props.history.push("/job_index")
 }
 
 componentWillMount(){
